refactor(imports): break circular dependencies in StorageClass and devices barrel
Break two circular dependency cycles that caused 'Class extends value
undefined' in tests and prevented using new Bean() in AI bean import.

Cycle 1 — StorageClass/UIHelper/UISettingsStorage:
- Move cloneData and getUnixTimestamp to StorageClass as static methods
- Remove static UIHelper import from StorageClass
- Use dynamic import for UIAlert in StorageClass and UIStorage to avoid
  indirect cycle through UIAlert → LoadingPopover → ShareService → UIHelper
- Delegate UIHelper.cloneData/getUnixTimestamp to StorageClass

Cycle 2 — devices barrel:
- Extract device enums to devices/types.ts
- Update 19 device files to import from ./types instead of ./index

With cycles broken:
- Remove karma-afterall-workaround (tests pass cleanly)
- Use new Bean() in bean-construction.service.ts instead of plain object
- Fix missing percentage field in createEmptyBeanInformation()
