+++ src/classes/devices/espressiScale.ts
@@ -1,6 +1,7 @@
 import { PeripheralData } from './ble.types';
 import { BluetoothScale, SCALE_TIMER_COMMAND, Weight } from './bluetoothDevice';
-import { Logger, ScaleType, sleep } from './index';
+import { Logger, sleep } from './common';
+import { ScaleType } from './types';
 
 declare var ble: any;
 export class EspressiScale extends BluetoothScale {
