+++ src/classes/devices/types.ts
@@ -1,0 +1,49 @@
+export enum BluetoothTypes {
+  SCALE = 'SCALE',
+  PRESSURE = 'PRESSURE',
+  TEMPERATURE = 'TEMPERATURE',
+  TDS = 'TDS',
+}
+
+export enum ScaleType {
+  DECENT = 'DECENT',
+  LUNAR = 'LUNAR',
+  JIMMY = 'JIMMY',
+  FELICITA = 'FELICITA',
+  FUTULA = 'FUTULA',
+  EUREKAPRECISA = 'EUREKAPRECISA',
+  SKALE = 'SKALE',
+  SMARTCHEF = 'SMARTCHEF',
+  DIFLUIDMICROBALANCE = 'DIFLUIDMIRCROBALANCE',
+  DIFLUIDMICROBALANCETI = 'DIFLUIDMIRCROBALANCETI',
+  BLACKCOFFEE = 'BLACKCOFFEE',
+  DIYPYTHONCOFFEESCALE = 'DIYPYTHONCOFFEESCALE',
+  DIYRUSTCOFFEESCALE = 'DIYRUSTCOFFEESCALE',
+  BOKOOSCALE = 'BOOKOOSCALE',
+  TIMEMORESCALE = 'TIMEMORESCALE',
+  VARIA_AKU = 'VARIA_AKU',
+  ESPRESSI = 'ESPRESSI',
+  WEIGHMYBRUSCALE = 'WEIGHMYBRUSCALE',
+}
+
+export enum PressureType {
+  POPSICLE = 'POPSICLE',
+  DIRECT = 'DIRECT',
+  PRS = 'PRS',
+  BOKOOPRESSURE = 'BOKOOPRESSURE',
+  COFFEESENSOR = 'COFFEESENSOR',
+}
+
+export enum TemperatureType {
+  ETI = 'ETI',
+  BASICGRILL = 'BASICGRILL',
+  MEATER = 'MEATER',
+  COMBUSTION = 'COMBUSTION',
+  ARGOS = 'ARGOS',
+  GEISINGER = 'GEISINGER',
+  COFFEESENSOR = 'COFFEESENSOR',
+}
+
+export enum RefractometerType {
+  R2 = 'R2',
+}
