--- src/karma-afterall-workaround.js
@@ -1,55 +1,0 @@
-// ---------------------------------------------------------------------------
-// Karma plugin: Ignore afterAll errors when all specs passed.
-//
-// Problem:
-//   The app has a circular import chain:
-//     StorageClass → UIHelper → UISettingsStorage → extends StorageClass
-//
-//   During Karma's afterAll teardown, a lazily-loaded chunk evaluates
-//   UISettingsStorage at a point where StorageClass is still undefined,
-//   throwing: "Class extends value undefined is not a constructor or null"
-//
-//   Karma-jasmine reports this as an afterAll ERROR, causing exit code 1
-//   even though every individual test spec passed.
-//
-// Solution:
-//   This custom reporter tracks actual spec results. After the run, if all
-//   specs passed and the only error is the known afterAll issue, it forces
-//   the exit code to 0.
-//
-// TODO: Remove this once the circular dependency in StorageClass is resolved.
-// ---------------------------------------------------------------------------
-
-function AfterAllWorkaroundReporter(baseReporterDecorator) {
-  baseReporterDecorator(this);
-
-  var specsFailed = 0;
-  var specsTotal = 0;
-
-  this.onSpecComplete = function (_browser, result) {
-    specsTotal++;
-    if (!result.success) {
-      specsFailed++;
-    }
-  };
-
-  this.onRunComplete = function (_browsers, results) {
-    if (specsFailed === 0 && specsTotal > 0 && results.error) {
-      // All specs passed but Karma flagged an error (afterAll teardown issue).
-      // Override the result so Karma exits with code 0.
-      results.exitCode = 0;
-      results.error = false;
-      console.log(
-        '\n[afterall-workaround] All ' +
-          specsTotal +
-          ' specs passed. Ignoring afterAll teardown error (known circular dependency).\n',
-      );
-    }
-  };
-}
-
-AfterAllWorkaroundReporter.$inject = ['baseReporterDecorator'];
-
-module.exports = {
-  'reporter:afterall-workaround': ['type', AfterAllWorkaroundReporter],
-};
